Rename routes files
